Clean up the two ugly bits called out in the journal
1. Extract FloatingOfflineChip from _layout.tsx into its own export in
   OfflineChip.tsx. The wrapper had no business living in the root layout —
   the layout's job is fonts/db/theming, not feature-specific overlay
   positioning. Moving it next to OfflineChip itself keeps the safe-area
   layout concerns colocated with the chip they position.

2. Drop the registerExtraProtectedHashes side-effect pattern. The pinning
   manager now imports pinnedFeedItemHashes directly and unions it into
   pinnedHashes() at call time. Structural guarantee replaces install-time
   ordering footgun — there is no longer a way to forget to register and
   silently lose pinned media to GC. The pinning test mocks feedItems the
   same way it already mocks @/content/store and preferences, so the RN
   import chain stays out of the unit-test path.
